# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ota/upload": {
      "post": {
        "operationId": "OtaController_uploadFirmware",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Subir un archivo binario de firmware (Solo Admin)",
        "tags": [
          "OTA"
        ]
      }
    },
    "/ota": {
      "post": {
        "operationId": "OtaController_createOtaUpdate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOtaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comando OTA publicado exitosamente."
          },
          "400": {
            "description": "Datos inválidos."
          },
          "403": {
            "description": "Acceso prohibido."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enviar actualización OTA de firmware a dispositivos (Solo Admin)",
        "tags": [
          "OTA"
        ]
      }
    },
    "/ota/history": {
      "get": {
        "operationId": "OtaController_getHistory",
        "parameters": [
          {
            "name": "deviceId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historial de actualizaciones OTA."
          },
          "403": {
            "description": "Acceso prohibido."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar historial de actualizaciones OTA (Solo Admin)",
        "tags": [
          "OTA"
        ]
      }
    },
    "/ota/available-files": {
      "get": {
        "operationId": "OtaController_getAvailableFiles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de archivos .bin."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar archivos de firmware disponibles (Solo Admin)",
        "tags": [
          "OTA"
        ]
      }
    },
    "/ota/latest": {
      "get": {
        "operationId": "OtaController_getLatestUpdate",
        "parameters": [
          {
            "name": "deviceId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar la última actualización disponible",
        "tags": [
          "OTA"
        ]
      }
    }
  },
  "info": {
    "title": "IMOX Cloud API",
    "description": "Documentación de la API para el sistema IMOX Cloud IoT",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOtaDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "description": "Versión objetivo del firmware (ej: v2.1.0)"
          },
          "url": {
            "type": "string",
            "description": "URL de descarga del binario del firmware (HTTPS)"
          },
          "hash": {
            "type": "string",
            "description": "Hash MD5/SHA256 del binario para verificación de integridad"
          },
          "target": {
            "type": "object",
            "description": "Dispositivos objetivo. Use \"ALL\" para todos los dispositivos activos, o un array de IDs: [1, 2, 3]",
            "examples": {
              "all": {
                "value": "ALL"
              },
              "specific": {
                "value": [
                  1,
                  2,
                  3
                ]
              }
            }
          }
        },
        "required": [
          "version",
          "url",
          "target"
        ]
      }
    }
  }
}